Engine 1.1
Major changes to the Listener and NotificationRequest classes allow for a
simpler, java.awt.event.Event-based calling method. Notification requests
no longer specify whether the type or code should be sent as parameters in
the method invokation; rather, a KeyEvent, MouseEvent, or MouseWheelEvent
is sent automatically depending on the type of event. These Events are
more useful than Integers type and code primarily because they supply both
the type and code as well as the point of origin for mouse presses,
releases, and moves.