Engine 1.2
Altered the ImageLoader so that there is an option to not sort the entries
lexicographically by refernce name. This makes indexes of specific images
found via getIndex() or add() reliable (if the list is ordered, they are
likely to change as new entries are added).
The Listener now has the ability to listen for mouse drags independently
of mouse movements. Thus, if a class listens only for mouse movements,
cases in which the mouse is held and then moved will be ignored.